requirements.txt
numpy
pandas
apscheduler
python-dotenv

